Fix iOS preview/capture mismatch + TFLite sample build, bump to JDK 17
- Preview now matches captured photo: iOS uses AVLayerVideoGravityResizeAspect
  (letterbox) and desktop ImagePanel scales-to-fit instead of stretching (#119).
- TFLite sample: move kflite off commonMain (no desktop artifact) into an
  expect/actual with a shared tfliteMain (android+iOS) and a desktop no-op;
  re-apply default hierarchy template so iOS actuals resolve.
- runTFliteModel is a suspend fun on Dispatchers.Default (was @Composable,
  which crashed when called from LaunchedEffect).
- Add iosApp Podfile (TensorFlowLiteObjC + Metal + CoreML) so kflite links;
  commit generated xcworkspace.
- Bump all module JVM toolchains 11 -> 17.
