+++ qrScannerPlugin/build.gradle.kts
@@ -14,7 +14,7 @@ group = "com.kashif.qr_scanner_plugin"
 version = "0.4"
 
 kotlin {
-    jvmToolchain(11)
+    jvmToolchain(17)
     androidTarget {
         publishLibraryVariants("release", "debug")
     }
